move sketch package under cli
