+++ cli/sketch/new.go
