+++ cli/sketch/sketch.go
